Change vBg function
Using the vBg function which is currently #vBg instead of the vBg function currently active under the section 'Standardize'. This way unit.temp and unit.pres are not defined in the stdVol call, but instead dependent on the value of unit.temp and unit.pres. These units are converted to optional units if any, in the stdVol function.