Arguments for mixed data
Codes for sorting were creating the cumBgDataPrep. Data is sorted as interval and standardized in the command under data sorting of mixed data. Therefore these are set as TRUE before entering cumBgVol. The code should be removed from cumBgVol(), but is currently kept in order for us to discuss in meeting.